fix git diff --stat
previously it would show changes between the current HEAD and the previous HEAD, which would neglect to show all the file changes from multiple commits, only the most recent one. this makes sure we store the HEAD hash from before the system is updated, and then compared to HEAD after it's updated.

Co-Authored-By: lloydbayley <21309308+lloydbayley@users.noreply.github.com>