Fix Excluded Species search for Chrome
Appears to be a issue scrolling a select list option into view and then setting it's selected state to true (as if the user had clicked on it) this seems to break all future calls of scrollIntoView on any options, regardless if the selected state if removed from all options.

scrollIntoView works perfectly otherwise and opted to apply a style to visually highlight the option. This turned out way better